Extract common abstract base
