+++ mc-java-entity/src/main/kotlin/io/spine/tools/mc/java/entity/EntityStateRenderer.kt
@@ -1,0 +1,81 @@
+/*
+ * Copyright 2024, TeamDev. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.tools.mc.java.entity
+
+import io.spine.protodata.MessageType
+import io.spine.protodata.java.JavaRenderer
+import io.spine.protodata.java.file.hasJavaOutput
+import io.spine.protodata.renderer.SourceFile
+import io.spine.protodata.renderer.SourceFileSet
+import io.spine.protodata.settings.loadSettings
+import io.spine.tools.mc.entity.DiscoveredEntities
+import io.spine.tools.mc.entity.EntityPluginComponent
+import io.spine.tools.mc.java.codegen.Entities
+
+/**
+ * An abstract state for renderers that generate the code for message types
+ * declared as [Entity][io.spine.base.EntityState]
+ *
+ * @see io.spine.tools.mc.entity.EntityDiscoveryProcess
+ */
+internal abstract class EntityStateRenderer : JavaRenderer(), EntityPluginComponent {
+
+    protected val settings: Entities by lazy {
+        loadSettings()
+    }
+
+    override fun render(sources: SourceFileSet) {
+        val relevant = sources.hasJavaOutput && settings.generateQueries
+        if (!relevant) {
+            return
+        }
+        val entityStates = foundEntityStates()
+        entityStates.forEach {
+            doRender(it, sources)
+        }
+    }
+
+    abstract fun doRender(type: MessageType, sources: SourceFileSet)
+
+    /**
+     * Obtains entity state types declared in all proto files.
+     */
+    private fun foundEntityStates(): List<MessageType> {
+        val discoveredEntities = select<DiscoveredEntities>().all()
+        val result = discoveredEntities.flatMap { it.typeList }
+        return result
+    }
+
+    protected fun SourceFileSet.fileOf(msg: MessageType): SourceFile {
+        val javaFile = javaFileOf(type = msg.name, declaredIn = msg.file)
+        val sourceFile = find(javaFile)
+        check(sourceFile != null) {
+            "Unable to locate the file `$sourceFile` in the source set `$this`."
+        }
+        return sourceFile
+    }
+}
